Fix CI pipeline failures: rename upload app and remove unused import
Rename `document_upload_app/app.py` to `document_upload_app/upload_app.py` to prevent shadowing the main `app` package during test collection.
Update `document_upload_app/test_app.py` to import from the renamed file.
Remove unused import `NewsArticle` from `tests/unit/test_news_service.py` to fix linting errors.
