+++ document_upload_app/test_app.py
@@ -1,7 +1,7 @@
 """Tests for the document upload Flask app."""
 import pytest
 from datetime import datetime
-from app import app, format_filename
+from upload_app import app, format_filename
 
 
 @pytest.fixture
