+++ tests/unit/test_news_service.py
@@ -6,7 +6,6 @@ import pytest
 from datetime import datetime
 from app.business.news_service import NewsService, ValidationError
 from app.data.repositories.news_repository import InMemoryNewsRepository
-from app.data.models.news_article import NewsArticle
 
 
 class TestNewsServiceValidation:
